app_flask/requirements.txt
cachelib==0.4.1
click==8.0.3
colorama==0.4.4
Flask==2.0.2
Flask-Session==0.4.0
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mysql-connector-python-rf==2.2.2
pytz==2021.3
Werkzeug==2.0.2
